프롬프트 객체 전달 수정 및 roomNumber 타입 Long으로 수정 (#13)
* feat: Room 객체 반환 함수 추가

* feat: 스케줄 페이지 관리 추가

* feat: 프롬프트 간 정보 교환

* edit: roomNumber Long 타입으로 수정

---------

+++ src/main/entity/MovieSchedule.java
@@ -14,12 +14,12 @@ public class MovieSchedule extends Entity<String> implements EntityValidator{
     private final LocalTime startAtTime;
     private final Room room;
 
-    public MovieSchedule(String id, Movie movie, LocalDate startAtDate, LocalTime startAtTime, int roomNumber) {
+    public MovieSchedule(String id, Movie movie, LocalDate startAtDate, LocalTime startAtTime, Long roomNumber) {
         super(id);
         this.movie = movie;
         this.startAtDate = startAtDate;
         this.startAtTime = startAtTime;
-        this.room = new Room((Room) DatabaseContext.getDatabase(Room.class).get(Integer.toString(roomNumber)));
+        this.room = new Room((Room) DatabaseContext.getDatabase(Room.class).get(Long.toString(roomNumber)));
     }
 
     public Movie getMovie() {
